refactor: 부동 소수점 문제를 해결하기 위한 타입 변경
- 기존 int 타입을 BigDecimal로 객체로 타입 변경

+++ src/main/java/blackjack/controller/BlackjackController.java
@@ -9,6 +9,7 @@ import blackjack.service.GameService;
 import blackjack.util.InputParser;
 import blackjack.view.InputView;
 import blackjack.view.OutputView;
+import java.math.BigDecimal;
 import java.util.List;
 
 public class BlackjackController {
@@ -53,7 +54,7 @@ public class BlackjackController {
 
     private BettingAmount readBetting(String name) {
         String input = inputView.readBettingAmount(name);
-        int amount = InputParser.parseBettingAmount(input);
+        BigDecimal amount = InputParser.parseBettingAmount(input);
         return new BettingAmount(amount);
     }
